Restore TypeForm module without eager union filtering
